requirements.txt
python-telegram-bot[job-queue]==20.7
web3==6.15.1
python-dotenv==1.0.1
requests==2.31.0
eth-account==0.11.0
eth-typing==4.0.0
eth-utils==2.3.1 